isleapp: Show error dialog if NOCD.si or ISLE.si fails to load (#95)
* isleapp: Show error dialog if NOCD.si fails to load

* isleapp: Return bool in IsleApp::Tick

foxtacles recommended me to return a boolean for if tick failed
instead of just setting `g_closed` and setting `g_closeResult`
to set the return value of the program exit.

I also think this is the better solution. since a similar solution
was used in LoadConfig.

* isleapp: Better error logs & return false in places

foxtacles recommended to use SDL's logging api to show detailed errors
and using false in two places where it should exit because of failure.

This commit adds these two things.

* isleapp: Add missing error dialog when ISLE.si fails

* isleapp: Add missing error dialog when ISLE.si fails

* isleapp: Remove weird duplicate message box

* Move error dialog to loop

---------

Co-authored-by: Christian Semmler <mail@csemmler.com>