Fix integer overflow in `LegoAnimationManager::Tickle`
This PR fixes a integer overflow in the `rand()` call inside of `LegoAnimationManager::Tickle`.
During animation sequences, like the one that plays when you first get in-game and the Infomaniac tells you about the Pizzeria, the `rand` function could return a massive value that would overflow the `int` type of `MxLong`.

Signed-off-by: kernaltrap <kernaltrap@gmail.com>