Add tooltips, remove obsolete directory stuff
In the config tool, the "Media Path" entry was removed, as it doesn't do anything in isle-portable as far as I know.
The "Disk Path" and "CD Path" options were merged into one directory editing thing, where CD Path is specified and Disk Path is then derived from that.
A checkbox to enable or disable full-screen was added.
Tooltips were added to the config app, so now users can more easily tell what a setting in the config tool does.