Revamp config tool, add tooltips and consolidate options. (#427)
* Add tooltips, remove obsolete directory stuff

In the config tool, the "Media Path" entry was removed, as it doesn't do anything in isle-portable as far as I know.
The "Disk Path" and "CD Path" options were merged into one directory editing thing, where CD Path is specified and Disk Path is then derived from that.
A checkbox to enable or disable full-screen was added.
Tooltips were added to the config app, so now users can more easily tell what a setting in the config tool does.

* Make clang-format happy

* oops that's an extra semicolon I don't need

* Remove obsolete full screen forcing

Co-authored-by: Anders Jenbo <anders@jenbo.dk>

* Update CONFIG/config.cpp

Co-authored-by: Anders Jenbo <anders@jenbo.dk>

---------

Co-authored-by: Anders Jenbo <anders@jenbo.dk>