terrible horrible no good very bad ""fixes""
The NXDK seems to be very very broken. Currently, these changes get things all the way to 100% compilation, but fail on the linking step due to `miniaudio`. These changes really need to be actually fixed in future to not be as bad as they are, but I haven't got much of a clue how best to do that.